# Speed up unit tests by adding default parallelism