--- a/requirements-dev.txt
+++ b/requirements-dev.txt
@@ -5,7 +5,8 @@ m2r2
 pluggy>=0.12.0
 pre-commit
 pylint<2.9
-pytest<3.7.0
+pytest==6.0.0
+pytest-xdist==2.2.1
 setuptools
 sphinx_rtd_theme
 tox==3.23.1