--- a/tox.ini
+++ b/tox.ini
@@ -5,7 +5,7 @@ envlist = py39,docs
 [testenv]
 deps = -rrequirements-dev.txt
 commands =
-    coverage run --source=elastalert/,tests/ -m pytest --strict {posargs}
+    coverage run --source=elastalert/,tests/ -m pytest --strict -n 4 {posargs}
     coverage report -m
     flake8 .