- Also update pytest to a newer version, making other changes required
  to enable this
- On a six-core machine, the test suite now runs in 8s as opposed to ~40s
- On a single-core machine, the test suite now runs in ~60s as opposed to ~85s
- Have also fixed one incorrect use of `raises()` which has been deprecated